feat(ivtt): use exact times, not ranges
Every commercial IVTT protocol I've found gives the incubation time as a
range (e.g. 2-4h), so at first I just copied these ranges verbatim into
the presets.  But eventually I decided that the philosophy behind
stepwise is to specify all of the exact details needed to perfrom the
protocol, and specifying time ranges goes against that.  So I replaced
all the ranges with the minimum recommended time, which is what I
normally use.  The end user of course still has the ability to make new
presets with different times, or to specify different times via the
command line.